Moved shortcut_recovery to GraphReader temporarily Added test to compare cache lookup to on-the-fly approach
