+++ test/test.cc
@@ -250,6 +250,8 @@ boost::property_tree::ptree make_config(const std::string& path_prefix,
         "max_concurrent_reader_users": 1,
         "reclassify_links": true,
         "shortcuts": true,
+        "shortcut_to_edge_cache": false,
+        "edge_to_shortcut_cache": false,
         "tile_dir": "%%",
         "tile_extract": "",
         "timezone": "%%/tz_world.sqlite",
